auth route

+++ server/api/dto/responses/auth_responses.go
@@ -1,12 +1,12 @@
 package responses
 
 type AuthLoginResponse struct {
-	Id    uint   `json:"id"`
+	Id    int32  `json:"id"`
 	Email string `json:"email"`
 	Token string `json:"token"`
 }
 
 type AuthVerifyResponse struct {
-	Id    float64 `json:"id"`
-	Email string  `json:"email"`
+	Id    int32  `json:"id"`
+	Email string `json:"email"`
 }
